refactor(plugin): move go-plugin adapter into package
Move the shared HashiCorp go-plugin handshake and gRPC adapter code into plugin/adapter so the supervisor owns the host-side plugin map separately. Update supervisor wiring to use the adapter package.
